+++ proofs/adapter/sql/showCreateTable.php
@@ -47,4 +47,32 @@ return static function() {
                 ->same($queries);
         },
     )->tag(Storage::sql);
+
+    yield test(
+        'Aggregate table name can be changed',
+        static function($assert) {
+            $show = ShowCreateTable::of(
+                Aggregates::of(Types::of(
+                    Type\PointInTimeType::of(new Clock),
+                ))->mapName(static fn($string) => match ($string) {
+                    User::class => 'some_user',
+                }),
+            );
+
+            // the match above allows to make sure the map is only applied to
+            // aggregate classes and not entities or pproperties
+
+            $queries = $show(User::class)
+                ->map(static fn($query) => $query->sql(Driver::mysql))
+                ->toList();
+
+            $assert
+                ->expected(
+                    <<<SQL
+                    CREATE TABLE  `some_user` (`id` char(36) NOT NULL  COMMENT 'UUID', `createdAt` char(32) NOT NULL  COMMENT 'Date with timezone down to the microsecond', `name` longtext  DEFAULT NULL COMMENT 'TODO adjust the type depending on your use case', `nameStr` longtext  DEFAULT NULL COMMENT 'TODO adjust the type depending on your use case', `role` longtext  DEFAULT NULL COMMENT 'TODO adjust the type depending on your use case', PRIMARY KEY (`id`))
+                    SQL,
+                )
+                ->in($queries);
+        },
+    )->tag(Storage::sql);
 };