+++ src/Definition/Aggregates.php
@@ -6,15 +6,31 @@ namespace Formal\ORM\Definition;
 final class Aggregates
 {
     private Types $types;
+    /** @var ?callable(class-string): non-empty-string  */
+    private $mapName;
 
-    private function __construct(Types $types)
+    /**
+     * @param callable(class-string): non-empty-string $mapName
+     */
+    private function __construct(Types $types, ?callable $mapName)
     {
         $this->types = $types;
+        $this->mapName = $mapName;
     }
 
     public static function of(Types $types): self
     {
-        return new self($types);
+        return new self($types, null);
+    }
+
+    /**
+     * @psalm-mutation-free
+     *
+     * @param callable(class-string): non-empty-string $map
+     */
+    public function mapName(callable $map): self
+    {
+        return new self($this->types, $map);
     }
 
     /**
@@ -26,6 +42,10 @@ final class Aggregates
      */
     public function get(string $class): Aggregate
     {
-        return Aggregate::of($this->types, $class);
+        return Aggregate::of(
+            $this->types,
+            $this->mapName,
+            $class,
+        );
     }
 }